Renamed shared methods module
